pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.23.0",
    "beautifulsoup4>=4.13.4",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.2",
    "openai>=1.84.0",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]
